// seller_backend/types/openapi.d.ts
export interface components {
    schemas: {
        Address: {
            name: string;
            line_one: string;
            line_two?: string;
            city: string;
            state: string;
            country: string;
            postal_code: string;
        };
        Buyer: {
            first_name: string;
            last_name: string;
            /** Format: email */
            email: string;
            phone_number?: string;
        };
        Item: {
            id: string;
            quantity: number;
        };
        PaymentProvider: {
            /** @enum {string} */
            provider: "stripe";
            supported_payment_methods: "card"[];
        };
        LineItem: {
            id: string;
            item: components["schemas"]["Item"];
            base_amount: number;
            discount: number;
            subtotal: number;
            tax: number;
            total: number;
        };
        Total: {
            /** @enum {string} */
            type: "items_base_amount" | "items_discount" | "subtotal" | "discount" | "fulfillment" | "tax" | "fee" | "total";
            display_text: string;
            amount: number;
        };
        FulfillmentOptionShipping: {
            /** @constant */
            type: "shipping";
            id: string;
            title: string;
            subtitle?: string;
            carrier?: string;
            /** Format: date-time */
            earliest_delivery_time?: string;
            /** Format: date-time */
            latest_delivery_time?: string;
            subtotal: string;
            tax: string;
            total: string;
        };
// ...
    };
    responses: never;
    parameters: {
        Authorization: string;
        ContentType: string;
        AcceptLanguage: string;
        UserAgent: string;
        IdempotencyKey: string;
        RequestId: string;
        Signature: string;
        Timestamp: string;
        APIVersion: string;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}